Gimbal Controller rework, use gimbal_device_id instead of compid, see long notes:
Implementation was based on the assumption that all the gimbal managers would
have a different compId, which is not always the case. As per mavlink documentation:

https://mavlink.io/en/services/gimbal_v2.html

Discovery should be based on GCS sending a MAV_CMD_REQUEST_MESSAGE for GIMBAL_MANAGER_INFORMATION
and we should instantiate gimbals when we receive response to this message.

This commit introduces several modifications toi allow this.

This was needed because otherwise the existent implementation would not allow multi gimbal
systems with gimbals belonging to the same compid to work correctly